emotion-detection/requirements.txt
fastapi
uvicorn
tensorflow-cpu
pillow
numpy
opencv-python-headless
python-multipart